drones-ingester/requirements.txt
pymongo==4.6.1
Flask==3.0.1
kafka-python==2.0.2